# sass/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MIS API 文档",
        "description": "伟明环保设备有限公司SIS系统 API 文档 api/documentation  php artisan l5-swagger:generate",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/auth/refresh": {
            "get": {
                "tags": [
                    "权限认证auth"
                ],
                "summary": "刷新token",
                "description": "使用说明：刷新token",
                "operationId": "refresh",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/dcs-standard/lists": {
            "get": {
                "tags": [
                    "DCS标准命名dcs-standard"
                ],
                "summary": "获取所有名称列表",
                "description": "使用说明：获取所有名称列表",
                "operationId": "dcs-standard-lists",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cn_name",
                        "in": "query",
                        "description": "TAG关键字搜索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_name",
                        "in": "query",
                        "description": "分组名搜索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "错误代码，0：为没有错误",
                                            "type": "integer",
                                            "default": "0"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DcsStandard"
                                            }
                                        },
                                        "message": {
                                            "description": "错误消息",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DcsStandard": {
                "title": "DcsStandard model",
                "description": "DcsStandard model",
                "properties": {
                    "id": {
                        "description": "创建数据模型",
                        "type": "integer"
                    },
                    "en_name": {
                        "type": "string"
                    },
                    "cn_name": {
                        "type": "string"
                    },
                    "sort": {
                        "type": "integer"
                    },
                    "is_show": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}